Time: 85 ms (41.21%), Space: 78.2 MB (35.61%) - LeetHub

+++ 1014-k-closest-points-to-origin/1014-k-closest-points-to-origin.cpp
@@ -1,10 +1,10 @@
 class Solution {
 public:
     vector<vector<int>> kClosest(vector<vector<int>>& points, int k) {
-        priority_queue<pair<long long,vector<int>>> pq;
+        priority_queue<pair<int,vector<int>>> pq;
         int n = points.size();
         for(int i = 0;i < n; i++){
-            long long distance = points[i][0]*points[i][0] + points[i][1]*points[i][1];
+            int distance = points[i][0]*points[i][0] + points[i][1]*points[i][1];
             pq.push({distance,points[i]});
             if(pq.size() > k){
                 pq.pop();
